# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API pour le jeu de quiz swipe",
        "title": "Quiz Swipe API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/question": {
            "get": {
                "description": "Retourne une question aléatoire du quiz sans la réponse correcte",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quiz"
                ],
                "summary": "Récupérer une question aléatoire",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.QuestionResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/swipe": {
            "post": {
                "description": "Vérifie si la direction du swipe correspond à la bonne réponse",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quiz"
                ],
                "summary": "Vérifier un swipe",
                "parameters": [
                    {
                        "description": "Swipe request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SwipeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SwipeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Vérifie que le serveur est en ligne",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.HealthResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Erreur lors de la récupération"
                }
            }
        },
        "models.HealthResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
        "models.QuestionResponse": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string",
                    "example": "animaux"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "image_url": {
                    "type": "string",
                    "example": "https://images.unsplash.com/photo-1548550023-2bdb3c5beed7"
                },
                "question": {
                    "type": "string",
                    "example": "Poule ou Coq ?"
                }
            }
        },
        "models.SwipeRequest": {
            "type": "object",
            "required": [
                "direction",
                "question_id"
            ],
            "properties": {
                "direction": {
                    "type": "string",
                    "enum": [
                        "left",
                        "right"
                    ],
                    "example": "right"
                },
                "question_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "models.SwipeResponse": {
            "type": "object",
            "properties": {
                "correct": {
                    "type": "boolean",
                    "example": true
                },
                "message": {
                    "type": "string",
                    "example": "Bravo! 🎉"
                }
            }
        }
    }
}